AVPlayerVC: volumeSlider done
